+++ loop_functions/kheperaiv_orca_failure_mqp_loop/mqp_http_client.cpp
@@ -53,7 +53,7 @@ bool mqp_http_client::solve(std::vector<std::vector<std::vector<std::vector<doub
                             std::string mode) {
 
     std::cout << "Calling the initial solve endpoint...\n" << std::endl;
-    std::string req_url = host+"/solve?k="+std::to_string(k)+"&n_a=" + std::to_string(int(n_a)) + "&fcr=" + std::to_string(fcr) +  "&ssd=" + std::to_string(int(ssd)) + "&mode=" + mode + "&rp=" + std::to_string(int(rp));
+    std::string req_url = host+"/solve?k="+std::to_string(k)+"&n_a=" + std::to_string(n_a) + "&fcr=" + std::to_string(fcr) +  "&ssd=" + std::to_string(ssd) + "&mode=" + mode + "&rp=" + std::to_string(rp);
 
     json data;
     mqp_packets::res mqp_res;
@@ -70,21 +70,21 @@ bool mqp_http_client::solve(std::vector<std::vector<std::vector<std::vector<doub
 }
 
 bool mqp_http_client::recalculate(std::vector<std::vector<std::vector<std::vector<double>>>> *path_arr,
-                            std::string host,
-                            int k,
-                            double n_a,
-                            double fcr,
-                            double rp,
-                            double ssd,
-                            std::string mode,
-                            std::string curr_fuel_levels,
-                            std::string curr_robots_pos
+                                  std::string host,
+                                  unsigned int k,
+                                  unsigned int n_a,
+                                  double fcr,
+                                  unsigned int rp,
+                                  double ssd,
+                                  std::string mode,
+                                  std::string curr_fuel_levels,
+                                  std::string curr_robots_pos
                           ) {
 
 
     std::cout << "Attempting to recalculate...\n" << std::endl;
 
-    std::string req_url = host+"/recalculate?job_id="+std::to_string(k)+"_" + std::to_string(int(n_a))  +  "_" + std::to_string(int(ssd)) + "_" + std::to_string(fcr) + "_" + std::to_string(int(rp)) + "_m&curr_robots_pos=" + curr_robots_pos + "&curr_fuel_levels=" + curr_fuel_levels;
+    std::string req_url = host+"/recalculate?job_id="+std::to_string(k)+"_" + std::to_string(n_a)  +  "_" + std::to_string(ssd) + "_" + std::to_string(fcr) + "_" + std::to_string(rp) + "_m&curr_robots_pos=" + curr_robots_pos + "&curr_fuel_levels=" + curr_fuel_levels;
     req_url = replace_all(req_url, " ", "");
     req_url = replace_all(req_url, ",", "%2C");
